+++ src/Enzo/_enzo.hpp
@@ -214,9 +214,9 @@ extern "C" {
 #include "amr/EnzoRefineMass.hpp"
 
 // [order dependencies:]
-#include "enzo_EnzoSinkParticle.hpp"
-#include "enzo_EnzoBondiHoyleSinkParticle.hpp"
-#include "enzo_EnzoFluxSinkParticle.hpp"
+#include "feedback-particle/EnzoSinkParticle.hpp"
+#include "feedback-particle/EnzoBondiHoyleSinkParticle.hpp"
+#include "feedback-particle/EnzoFluxSinkParticle.hpp"
 
 // [order dependencies:]
 #include "enzo_EnzoEFltArrayMap.hpp"
@@ -245,27 +245,27 @@ extern "C" {
 #include "hydro-mhd/EnzoBfieldMethod.hpp"
 #include "hydro-mhd/EnzoBfieldMethodCT.hpp"
 
-#include "enzo_EnzoMethodAccretion.hpp"
+#include "feedback-particle/EnzoMethodAccretion.hpp"
 #include "enzo_EnzoMethodBackgroundAcceleration.hpp"
-#include "enzo_EnzoMethodBondiHoyleAccretion.hpp"
+#include "feedback-particle/EnzoMethodBondiHoyleAccretion.hpp"
 #include "enzo_EnzoMethodCheck.hpp"
 #include "enzo_EnzoMethodComovingExpansion.hpp"
 #include "enzo_EnzoMethodCosmology.hpp"
-#include "enzo_EnzoMethodDistributedFeedback.hpp"
-#include "enzo_EnzoMethodFeedback.hpp"
-#include "enzo_EnzoMethodFeedbackSTARSS.hpp"
-#include "enzo_EnzoMethodFluxAccretion.hpp"
+#include "feedback-particle/EnzoMethodDistributedFeedback.hpp"
+#include "feedback-particle/EnzoMethodFeedback.hpp"
+#include "feedback-particle/EnzoMethodFeedbackSTARSS.hpp"
+#include "feedback-particle/EnzoMethodFluxAccretion.hpp"
 #include "enzo_EnzoMethodGrackle.hpp"
 #include "gravity/EnzoMethodGravity.hpp"
 #include "enzo_EnzoMethodHeat.hpp"
-#include "enzo_EnzoMethodMergeSinks.hpp"
+#include "feedback-particle/EnzoMethodMergeSinks.hpp"
 #include "enzo_EnzoMethodPmDeposit.hpp"
 #include "enzo_EnzoMethodPmUpdate.hpp"
-#include "enzo_EnzoMethodSinkMaker.hpp"
-#include "enzo_EnzoMethodStarMaker.hpp"
-#include "enzo_EnzoMethodStarMakerSTARSS.hpp"
-#include "enzo_EnzoMethodStarMakerStochasticSF.hpp"
-#include "enzo_EnzoMethodThresholdAccretion.hpp"
+#include "feedback-particle/EnzoMethodSinkMaker.hpp"
+#include "feedback-particle/EnzoMethodStarMaker.hpp"
+#include "feedback-particle/EnzoMethodStarMakerSTARSS.hpp"
+#include "feedback-particle/EnzoMethodStarMakerStochasticSF.hpp"
+#include "feedback-particle/EnzoMethodThresholdAccretion.hpp"
 #include "enzo_EnzoMethodTurbulence.hpp"
 #include "hydro-mhd/EnzoMethodHydro.hpp"
 #include "hydro-mhd/EnzoMethodMHDVlct.hpp"
